Updates for new traverse
Added a new find_zero_pitch code that has the option to select either of the traverses for alignment.

Corrected conversion for the new traverse (deg to volts). This resolves a compatibility issue from move_to_new_pos. This was not an issue for the run motors code because there the conversion from deg to volts is bypassed by the traversecmd function.

Moved out trial data.